pyproject.toml
[project]
name = "kinoweek"
version = "0.1.0"
description = "Astor Kino Notifier - Web scraping and Telegram notification system"
authors = [{name = "Sia"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "ics>=0.7.2",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kinoweek"]

[project.scripts]
kinoweek = "src.kinoweek.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "T20", "PT", "Q"]
ignore = ["E501", "PIE804"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
